user data loaders refactored
